Update TextField component and tests, add TextInput component
