+++ packages/design-system/src/components/TextField/TextField.test.jsx
@@ -1,6 +1,6 @@
-import TextField, { unmaskValue } from './TextField';
 import { mount, shallow } from 'enzyme';
 import React from 'react';
+import TextField from './TextField';
 
 function render(customProps = {}, deep = false) {
   const props = Object.assign(
@@ -19,77 +19,6 @@ function render(customProps = {}, deep = false) {
 }
 
 describe('TextField', function () {
-  it('is an input field', () => {
-    const data = render();
-    const field = data.wrapper.find('.ds-c-field').first();
-
-    expect(field.is('input')).toBe(true);
-    expect(field.prop('rows')).toBeUndefined();
-    expect(field.prop('type')).toBe('text');
-  });
-
-  it('is a textarea', () => {
-    const data = render({ multiline: true });
-    const field = data.wrapper.find('.ds-c-field').first();
-
-    expect(field.is('textarea')).toBe(true);
-    expect(field.prop('rows')).toBeUndefined();
-    expect(field.prop('type')).toBeUndefined();
-  });
-
-  it('is a password field', () => {
-    const data = render({ type: 'password' });
-    const field = data.wrapper.find('.ds-c-field').first();
-
-    expect(field.prop('type')).toBe('password');
-  });
-
-  it('is disabled', () => {
-    const data = render({ disabled: true });
-    const field = data.wrapper.find('.ds-c-field').first();
-
-    expect(field.prop('disabled')).toBe(data.props.disabled);
-  });
-
-  it('has a defaultValue', () => {
-    const data = render({ defaultValue: 'Yay' });
-    const field = data.wrapper.find('.ds-c-field').first();
-
-    expect(field.prop('defaultValue')).toBe(data.props.defaultValue);
-    expect(field.prop('value')).toBeUndefined();
-  });
-
-  it('has a value', () => {
-    const data = render({ value: 'Yay' });
-    const field = data.wrapper.find('.ds-c-field').first();
-
-    expect(field.prop('value')).toBe(data.props.value);
-    expect(field.prop('defaultValue')).toBeUndefined();
-  });
-
-  it('shows 5 rows of text', () => {
-    const data = render({
-      multiline: true,
-      rows: 5,
-    });
-    const field = data.wrapper.find('.ds-c-field').first();
-
-    expect(field.prop('rows')).toBe(data.props.rows);
-  });
-
-  it('has a unique id', () => {
-    const fieldData1 = render();
-    const fieldData2 = render();
-    const field1 = fieldData1.wrapper.find('.ds-c-field').first();
-    const label1 = fieldData1.wrapper.find('FormLabel').first();
-    const field2 = fieldData2.wrapper.find('.ds-c-field').first();
-    const idRegex = new RegExp('textfield_[0-9]+');
-
-    expect(field1.prop('id')).toMatch(idRegex);
-    expect(label1.prop('fieldId')).toBe(field1.prop('id'));
-    expect(field1.prop('id')).not.toBe(field2.prop('id'));
-  });
-
   it('has a label', () => {
     const data = render();
     const label = data.wrapper.find('FormLabel').first();
@@ -110,69 +39,6 @@ describe('TextField', function () {
     expect(data.wrapper.hasClass('bar')).toBe(true);
   });
 
-  it('adds className to field', () => {
-    const data = render({ fieldClassName: 'bar' });
-    const field = data.wrapper.find('.ds-c-field').first();
-
-    expect(field.hasClass('ds-c-field')).toBe(true);
-    expect(field.hasClass('bar')).toBe(true);
-  });
-
-  it('adds className to label', () => {
-    const data = render({ labelClassName: 'bar' });
-
-    expect(data.wrapper.find('FormLabel').hasClass('bar')).toBe(true);
-  });
-
-  it('adds size classes to input', () => {
-    const mediumData = render({ size: 'medium' });
-    const mediumField = mediumData.wrapper.find('.ds-c-field').first();
-    const smallData = render({ size: 'small' });
-    const smallField = smallData.wrapper.find('.ds-c-field').first();
-
-    expect(mediumField.hasClass('ds-c-field--medium')).toBe(true);
-    expect(smallField.hasClass('ds-c-field--small')).toBe(true);
-  });
-
-  it('adds min/max input attributes', () => {
-    const data = render({
-      max: 10,
-      min: 1,
-    });
-    const field = data.wrapper.find('.ds-c-field').first();
-
-    expect(field.prop('max')).toBe(data.props.max);
-    expect(field.prop('min')).toBe(data.props.min);
-  });
-
-  it('adds aria-label attribute', () => {
-    const data = render({
-      ariaLabel: 'Foo',
-    });
-    const field = data.wrapper.find('.ds-c-field').first();
-
-    expect(field.prop('aria-label')).toBe(data.props.ariaLabel);
-  });
-
-  it('adds overrides default aria-label with defined prop', () => {
-    const data = render({
-      ariaLabel: 'Foo',
-      mask: 'currency',
-    });
-    const field = data.wrapper.find('.ds-c-field').first();
-
-    expect(field.prop('aria-label')).toBe(data.props.ariaLabel);
-  });
-
-  it('adds undocumented prop to input field', () => {
-    const data = render({
-      'data-foo': 'bar',
-    });
-    const field = data.wrapper.find('.ds-c-field').first();
-
-    expect(field.prop('data-foo')).toBe(data.props['data-foo']);
-  });
-
   it('returns reference to input field', () => {
     let ref;
     const data = render(
@@ -186,20 +52,6 @@ describe('TextField', function () {
     );
 
     expect(ref.value).toBe(data.props.defaultValue);
-  });
-
-  it('focuses the input when focusTrigger is passed', () => {
-    const data = render(
-      {
-        id: 'focus',
-        focusTrigger: true,
-      },
-      true
-    );
-
-    setTimeout(() => {
-      expect(data.wrapper.find('input').props().id).toEqual(document.activeElement.id);
-    }, 20);
   });
 
   describe('has error', () => {
@@ -234,56 +86,17 @@ describe('TextField', function () {
     });
   });
 
-  describe('event handlers', () => {
-    let data;
+  it('focuses the input when focusTrigger is passed', () => {
+    const data = render(
+      {
+        id: 'focus',
+        focusTrigger: true,
+      },
+      true
+    );
 
-    beforeEach(() => {
-      data = render({
-        onBlur: jest.fn(),
-        onChange: jest.fn(),
-      });
-    });
-
-    it('calls onBlur', () => {
-      data.wrapper.find('.ds-c-field').first().simulate('blur');
-
-      expect(data.props.onBlur.mock.calls.length).toBe(1);
-    });
-
-    it('calls onChange', () => {
-      data.wrapper.find('.ds-c-field').first().simulate('change');
-
-      expect(data.props.onChange.mock.calls.length).toBe(1);
-    });
-  });
-
-  describe('masks', () => {
-    it('exports unmaskValue method', () => {
-      expect(typeof unmaskValue).toBe('function');
-    });
-
-    it('renders TextField', () => {
-      const data = render();
-
-      expect(data.wrapper).toMatchSnapshot();
-    });
-
-    it('renders currency mask', () => {
-      const data = render({
-        mask: 'currency',
-      });
-
-      expect(data.wrapper).toMatchSnapshot();
-    });
-
-    it('updates input classes when props are updated', () => {
-      const wrapper = render({ mask: 'currency' }, true).wrapper;
-
-      expect(wrapper.find('input').hasClass('ds-c-field--error')).toBe(false);
-
-      wrapper.setProps({ errorMessage: 'Oh no' });
-
-      expect(wrapper.find('input').hasClass('ds-c-field--error')).toBe(true);
-    });
+    setTimeout(() => {
+      expect(data.wrapper.find('input').props().id).toEqual(document.activeElement.id);
+    }, 20);
   });
 });
